requirements.txt
# Core dependencies
requests>=2.31.0           # HTTP client for LM Studio API
PyYAML>=6.0.1              # Config file parsing
python-dotenv>=1.0.0       # Environment variables

# Document processing
PyPDF2>=3.0.1              # PDF text extraction
python-docx>=1.1.0         # DOCX processing
python-pptx>=0.6.23        # PPTX processing
openpyxl>=3.1.2            # Excel file processing

# Vision models (MLX)
mlx>=0.4.0                 # Apple MLX framework
mlx-vlm>=0.0.9             # Vision language models for MLX

# Image processing
Pillow>=10.0.0             # Image handling

# Video processing
opencv-python>=4.8.0       # Video frame extraction

# Text processing
regex>=2023.10.3           # Advanced regex for text cleaning

# CLI and utilities
click>=8.1.7               # CLI framework
tqdm>=4.66.0               # Progress bars
colorama>=0.4.6            # Colored terminal output
tabulate>=0.9.0            # Table formatting for output

# Testing
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reports
pytest-mock>=3.12.0        # Mocking utilities

# Code quality
black>=23.12.0             # Code formatting
flake8>=6.1.0              # Linting
mypy>=1.7.0                # Type checking
